+++ test/integration/server/context/test-class-in-context.ts
@@ -1,0 +1,11 @@
+export class TestClassInContext {
+    public someProperty?: number;
+
+    constructor(somePropertyInitialValue: number) {
+        this.someProperty = somePropertyInitialValue;
+    }
+
+    public doSomething(): string {
+        return `did something successfully with someProperty of ${this.someProperty}`;
+    }
+}
